Version 1.1
Mainly bug fixes. New features:- Added the uri history://n to access the n-th newest item in the history.   Similarly, the :History command may now accecpt a numeric argument.- The last access time for each link now appears on the history page- Added the g:browser_connect_timeout variable to control when we give up   connecting to a remote site- Added support for homepage in the plugin. :Browse and similar commands can   now be called with no arguments.- Changed the loading of the plugin so that vim startup time is improved.- Added a command for showing the current mappings, and a command to extend   this list for user defined mappings